Fix signed/unsigned and size mismatch MSVC warnings
MSVC signed/unsigned comparison warnings were addressed. Typically by
aligning variable types, where int was used as the iterator and unsigned
was used for limiter.

32/64-bit size mismatches warnings were addressed. Typically by
migrating to size_t when up-sizing and type-casting when down-sizing.

+++ src/openvpn/block_dns.c
@@ -370,7 +370,7 @@ get_interface_metric(const NET_IFINDEX index, const ADDRESS_FAMILY family)
         }
         return ipiface.Metric;
     }
-    return -err;
+    return -(int)err;
 }
 
 /*
